+++ CHANGELOG.md
@@ -29,8 +29,8 @@ Releasing is documented in RELEASE.md
 
 ### Added
 - full support for encoded values ([#2108](https://github.com/GIScience/openrouteservice/issues/2108))
-- enable `ors_surface` encoded value ([#2110](https://github.com/GIScience/openrouteservice/issues/2110))
-- enable `way_type` encoded value ([#2115](https://github.com/GIScience/openrouteservice/issues/2115))
+- enable encoded value for way surface ([#2110](https://github.com/GIScience/openrouteservice/issues/2110))
+- enable encoded value for way type ([#2115](https://github.com/GIScience/openrouteservice/issues/2115))
 
 ### Changed
 - transition from extended storage to encoded values for storing way type and surface ([#2113](https://github.com/GIScience/openrouteservice/pull/2113), [#2115](https://github.com/GIScience/openrouteservice/pull/2115))
@@ -38,12 +38,20 @@ Releasing is documented in RELEASE.md
 ### Deprecated
 
 ### Removed
+- replace `WaySurfaceType` external storage by corresponding encoded values ([#2116](https://github.com/GIScience/openrouteservice/pull/2116))
 
 ### Fixed
 - remove spurious entry for an unassigned country ID value 137 from the documentation ([#2103](https://github.com/GIScience/openrouteservice/pull/2103))
 
 ### Security
-
+- update spring-boot-starter to 3.4.9 fixes 
+  [CVE-2025-48989](https://www.cve.org/CVERecord?id=CVE-2025-48989),
+  [CVE-2025-41242](https://www.cve.org/CVERecord?id=CVE-2025-41242),
+  [CVE-2025-48988](https://www.cve.org/CVERecord?id=CVE-2025-48988),
+  [CVE-2025-53506](https://www.cve.org/CVERecord?id=CVE-2025-53506),
+  [CVE-2025-52520](https://www.cve.org/CVERecord?id=CVE-2025-52520) and more
+- update apache commons-lang3 to 3.18.0 fixes 
+  [CVE-2025-48924](https://www.cve.org/CVERecord?id=CVE-2025-48924),
 
 ## [9.3.1] - 2025-07-28